Add ALT text to the generated <img> HTML element
* graphviz.py: Add the GRAPHVIZ_ALT_TEXT configuration variable.

* mdx_graphviz.py: Set the ALT property of the <img> element
according to this order of priority : (1) form the (new) block option
"alt-text", (2) from the ID of the Graphviz object, or (3) from the
GRAPHVIZ_ALT_TEXT configuration option.

In this version, the specification of key/value pairs has been improved. It
is now possible to use quotes around the value, like key="val". The
previous way of specifying key/value pairs (key=val) is still valid.

This changes fixes the issue #30.