Re-enable TestGraphvizAltTextWithoutID
This test was failing in GitHub Actions because ubuntu-latest currently
resolves to Ubuntu 24.04, which ships an ancient version of graphviz
containing a bug in which the Title: SVG comment is set to '%3' if no
title is specified in the input:

https://gitlab.com/graphviz/graphviz/-/issues/1376

This bug doesn't reproduce on my AlmaLinux 10 dev server (AlmaLinux uses
a much more recent version of graphviz), but I was able to reproduce it
locally on Ubuntu 24.04.

Since Ubuntu is very popular, though, it's probably worth implementing a
workaround in this plugin.  This commit proposes treating a title of
'%3' as unspecified.

Incidentally, this is the second bug I've encountered recently in which
a bug in an outdated package shipped in current Debian distros caused
problems and confusion:

https://github.com/wasm-bindgen/wasm-bindgen/issues/4528#issuecomment-3075221775