[tests] Give expected values, settings as setUp() arguments
This changes how input, settings, and expected output are specified in
test cases.  Internal logic within the TestGraphviz base class is
removed, in favor of specifying all three of these explicitly.  Input
options are now given as a dict.

The main goal of this change is to make it easier to extend these
tests to cover RST too.  IMO this also makes it easier to read the tests
and understand which combination of settings and options are under test,
and what the expected output is.

A noteworthy side-effect of this change is that we now test the plugin's
behavior in the case that no GRAPHVIZ_* settings are specificed, except
those explicitly overridden in test cases.