Factor out compressed img element logic
This factors out the logic used to create the base64-compressed img
element, which can be reused in the RST path.

Unfortunately, I didn't see a good way to similarly factor out the
non-compressed code.  The problem is that the Markdown path relies on
working with ElementTree, and there setting the raw SVG output as the
container's `text` apparently works; but in the later-added RST path,
setting the SVG as `text` causes it to be escaped when I then pass it to
ET.tostring.

(I think this could be solved by using lxml instead, but that would
require a bigger change to the existing Markdown path.)